requirements.txt
# -------------------------
# Core Backend
# -------------------------
fastapi==0.110.0
uvicorn[standard]==0.27.1

# -------------------------
# Data Processing
# -------------------------
pandas==2.2.1
numpy==1.26.4

# -------------------------
# Database (future-proof)
# -------------------------
psycopg2-binary==2.9.9

# -------------------------
# Email / Utils
# -------------------------
python-dotenv==1.0.1
email-validator==2.1.0

# -------------------------
# HTTP / API
# -------------------------
requests==2.31.0

# -------------------------
# Optional (safe to keep)
# -------------------------
python-multipart==0.0.9
Kaggle==1.5.16